database_examples/requirements.txt

# postgresql
psycopg2-binary

